snb/systemagent-r6: Remove stack check in GlueAllocatePool
When allocating the pool, the code bails out on a signed check
if (stack - pool) < 1000. The pool location is fixed at 0xff7e1000,
which places restrictions on the stack location (must be above the
heap). This binary patches the systemagent-r6 binary to not bail out,
by changing a "JG" to "JMP". This allows to put the stack wherever we
want as long as it does not collide.

Change-Id: If6b4094ba0208ba14dea0bb35119906dccc4fd38
Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>