systemagent-r6: Move the pool
Now that the check of the pool wrt the stack is gone, the pool can be
relocated from 0xff7e1000 to 0xff7f7000.

The advantage is that the pool and the MRC_VAR region are now next to
each other which allows coreboot to have one continuous CAR region
for both coreboot (DCACHE_RAM_SIZE) and the mrc.bin
(DCACHE_RAM_MRC_VAR_SIZE), which can remove the current check from the
linker script.

Change-Id: I8223bf6dae09b7c4accd2d23ec076600bbeb11f0
Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>