mainboard/facebook/fbg1701: Add descriptor and Intel ME
These descriptor has been taken from the vendor BIOS.
License.txt has been added.

The person or company downloading this software (cloning the repository) is allowed to use this
internally for development purposes and is bound by the restrictions imposed by this agreement.

We are allowed to put the binaries on the git as long as this agreement is included as well.
The blobs should not show up in a public git without this license and a coreboot image with these
blobs included should always include exhibit A. That is what this agreement tries to achieve.

The Intel stakeholders are aware of this and know how the blobs will be published in the 3rdparty
blobs directory.

BUG=N/A
TEST=booting Facebook FBG1701

Change-Id: Ic03da722bc30da48ce38560745a3f89f6bd37943
Signed-off-by: Frans Hendriks <fhendriks@eltan.com>