mb/google/guybrush: Update APCB file
The existing APCB file only has 4 SPD slots. This will not be enough
for DeWatt devices due to upcoming changes in how RAM IDs are allocated
for that variant. This commit updates the APCB file to have 16 slots
which is sufficient for DeWatt.

BUG=b:224884904
TEST=Used apcb_v3_edit to verify that the APCB file has 16 slots,
checked that AP firmware images built with this file boot correctly

Signed-off-by: Robert Zieba <robertzieba@google.com>
Change-Id: Ifbfe2c61c42cd503a70fd84c51ce184c40fed318