mb/google/skyrim: Add RO SPL table
ChromeOS requires a RO SPL table. Add it here so that it can be linked
in coreboot.

The SPL table contains a set of version numbers to prevent rollback
attacks. Updates with a value lower than the value in the table are
not allowed.

See the Versioned Chip Endorsement Key (VCEK) Certificate and KDS
Interface Specification.  Document # 57230 Rev. 0.50 October 2021
https://web.archive.org/web/20221213033802/https://www.amd.com/system/files/TechDocs/57230.pdf

BUG=b:243470283
TEST=Build Skyrim BIOS image and boot to OS.

Change-Id: Iee897dd2c0943c17e81e02a4d6c6296b585e12af
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>